+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Adds a detector for Discord message deletion, which is handled if it's a game.
+
 ## [v8.6.1](https://github.com/lexicalunit/spellbot/releases/tag/v8.6.1) - 2022-12-23
 
 ### Changed